consolidated URL helpers into simpler constants
