+++ libs/instagram/src/lib/constants/urls.ts
@@ -2,5 +2,9 @@
  * @description  Crawled URL's belonging to Instagram
  *               Broken apart into BASE & EXT (URL extensions)
  */
-export const INSTAGRAM_BASE_URL = 'https://www.instagram.com'
-export const INSTAGRAM_URL_EXT_LOGIN = '/accounts/login'
+export const INSTAGRAM_URL_BASE = 'https://www.instagram.com/'
+
+export const INSTAGRAM_URL_LOGIN = INSTAGRAM_URL_BASE + 'accounts/login/'
+export const INSTAGRAM_URL_MESSAGING = INSTAGRAM_URL_BASE + 'direct/inbox/'
+export const INSTAGRAM_URL_EXPLORE = INSTAGRAM_URL_BASE + 'explore/'
+export const INSTAGRAM_URL_SETTINGS = INSTAGRAM_URL_BASE + 'accounts/edit/'
